moved state back to app.tsx
