+++ Week-9/todolist/src/components/Todos.tsx
@@ -1,26 +1,15 @@
-import React, { useState } from 'react';
 import Todo from './Todo';
-import todoListData from '../data/todolist.json';
 
-function Todos() {
-  const [todoList, setTodoList] = useState(todoListData);
-
-  const handleToggle = (id: any) => {
-    let mapped = todoList.map((task) => {
-      return task.id == id
-        ? { ...task, complete: !task.complete }
-        : { ...task };
-    });
-    setTodoList(mapped);
-  };
+const Todos = (props: any) => {
+  const { todoList, handleToggle } = props;
 
   return (
     <>
-      {todoList.map((todo) => {
+      {todoList.map((todo: any) => {
         return <Todo key={todo.id} todo={todo} handleToggle={handleToggle} />;
       })}
     </>
   );
-}
+};
 
 export default Todos;
